+++ qt_ui/windows/mission/flight/payload/propertyspinbox.py
@@ -1,14 +1,16 @@
 from PySide6.QtWidgets import QSpinBox
 from dcs.unitpropertydescription import UnitPropertyDescription
 
-from game.ato import Flight
+from game.ato.flightmember import FlightMember
 from .missingpropertydataerror import MissingPropertyDataError
 
 
 class PropertySpinBox(QSpinBox):
-    def __init__(self, flight: Flight, prop: UnitPropertyDescription) -> None:
+    def __init__(
+        self, flight_member: FlightMember, prop: UnitPropertyDescription
+    ) -> None:
         super().__init__()
-        self.flight = flight
+        self.flight_member = flight_member
         self.prop = prop
 
         if prop.minimum is None:
@@ -20,9 +22,17 @@ class PropertySpinBox(QSpinBox):
 
         self.setMinimum(prop.minimum)
         self.setMaximum(prop.maximum)
-        self.setValue(self.flight.props.get(self.prop.identifier, self.prop.default))
+        self.setValue(
+            self.flight_member.properties.get(self.prop.identifier, self.prop.default)
+        )
 
         self.valueChanged.connect(self.on_value_changed)
 
     def on_value_changed(self, value: int) -> None:
-        self.flight.props[self.prop.identifier] = value
+        self.flight_member.properties[self.prop.identifier] = value
+
+    def set_flight_member(self, flight_member: FlightMember) -> None:
+        self.flight_member = flight_member
+        self.setValue(
+            self.flight_member.properties.get(self.prop.identifier, self.prop.default)
+        )